requirements.txt
openai-whisper>=20231117
openai>=1.0.0
PyYAML>=6.0
tqdm>=4.0.0
